requirements.txt
faster-whisper>=0.10.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pyautogui>=0.9.54
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
rumps>=0.4.0
